Moved functions into proper headers
