+++ VentilatorControl/src/alarms.h
@@ -1,0 +1,7 @@
+#ifndef __ALARMS_H__
+#define __ALARMS_H__
+
+void alarmsHandler (void);
+void alarmsVisualAudioHandler (void);
+
+#endif // __ALARMS_H__
