+++ VentilatorControl/src/compute_comms.h
@@ -3,6 +3,7 @@
 #ifndef __COMPUTE_COMMS_H__
 #define __COMPUTE_COMMS_H__
 
+void computeSerialSend (void);
 void computeSerialReceive (void);
 
 #endif // __COMPUTE_COMMS_H__
